requirements.txt
faiss-cpu
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
requests
faiss-cpu
numpy
redis
rq
tenacity
structlog
prometheus-client
gunicorn
flask